statsig: reduce package surface area
With this change, we mostly focus on removing the subpackages. We'd like
to unwind the various components a bit and doing so will be easier if we
can first remove some of the package relationships.

The effect is a clear init path for the client interface, a simplified
transport object and the ability to access internal fields between
components.

With this change, we can take it further by defining an interface
between the API based gate checks and the local evaluator.

Signed-off-by: Stephen Day <stephen.day@getcruise.com>